[sk] Refactored some templates and utils
