+++ mage_ai/data_preparation/templates/transformers/default.py
@@ -1,6 +1,15 @@
 from pandas import DataFrame
 
 
-def transform_df(df) -> DataFrame:
+def transform_df(df: DataFrame) -> DataFrame:
+    """
+    Template code for a transformer block.
+
+    Args:
+        df (DataFrame): Input data frame from previously executed block.
+
+    Returns:
+        DataFrame: Transformed data frame
+    """
     # Your logic here
     return df
